target/python/requirements.txt
# Required Python packages for Whisper transcription
openai-whisper==20231117
torch>=2.0.0
numpy>=1.20.0
ffmpeg-python>=0.2.0
setuptools>=65.5.1
tqdm>=4.66.0

# Note: ffmpeg is also required as a system dependency
# For Windows: Install from https://ffmpeg.org/download.html
# For macOS: brew install ffmpeg
# For Linux: apt-get install ffmpeg
